pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "email-validator>=2.3.0",
    "flask>=3.1.3",
    "flask-login>=0.6.3",
    "flask-sqlalchemy>=3.1.1",
    "gunicorn>=25.3.0",
    "pandas>=3.0.2",
    "psycopg2-binary>=2.9.12",
    "python-dotenv>=1.2.2",
    "sqlalchemy>=2.0.49",
    "werkzeug>=3.1.8",
]
